requirements.txt
discord.py==2.1.1
docutils==0.19
